Simple SPI config allows output of 32MHz into MCU
Created a simple SPI configuration that allows us to talk to the transceiver and then get it to give the MCU the FULL 32MHz signal. It is by default configured to divide it down to 1 MHz, probably for power saving reasons. However, we don't want that because the PLL can't be used when the MCU external clock reference is less than 2 MHz!!!