Clock configuration finished
The MCU now boots up with the internal oscillator like it should then sets up the port and spi clocks (so that we can read/write those registers) then a small initialization procedure on the SPI0 module requests the full 32MHz signal from the transceiver. When the transceiver supplies that, we shut off all the clocks to all the modules and then dial up the PLL to create a 48 MHz clock signal. After that, we turn the clocks back on for the modules we use in this code (which aren't much some are even completey unnecessary) then blink some LEDs because thats what we do in computer engineering is blink LEDs. Now lets get this thing space ready so we can blink LEDs out in space that no one can see. The end.