Added static flash config, removed process stack
The flash addresses 0x400-0x410 are used to, among other things, set the
security setting - disabling debugging. We had been running into not
being able to debug the chip when programs got over a certain length. It
turns out they were extending into this address range and setting the
security bit. Linker magic in a previous commit places this new magic
struct at the right location to correctly configure the flash options,
and then place the program after.

Also removed the process stack setup because we're only using the main
stack. Sometime in the future we may need it again, but not today.